🚧 wip
Former-commit-id: 961e3c39c1b0b14e4ac9809a60de4be1eaa91415
Former-commit-id: 13e2a4b63604020f6691422b0e04cdf957bbed31

+++ src/framework/meshblock.cpp
@@ -164,7 +164,7 @@ namespace ntt {
   }
 
   template <Dimension D>
-  auto Mesh<D>::rangeCells(const tuple_t<tuple_t<short, Dim2>, D>& ranges) -> range_t<D> {
+  auto Mesh<D>::rangeCells(const tuple_t<tuple_t<int, Dim2>, D>& ranges) -> range_t<D> {
     tuple_t<int, D> imin, imax;
     for (short i = 0; i < (short)D; i++) {
       if ((ranges[i][0] < -N_GHOSTS) || (ranges[i][1] > N_GHOSTS)) {
